+++ src/ndv/__init__.py
@@ -12,9 +12,9 @@ __email__ = "talley.lambert@example.com"
 from typing import TYPE_CHECKING
 
 from . import data
+from ._old_viewer import NDViewer
+from .models._data_wrapper import DataWrapper
 from .util import imshow
-from .viewer._data_wrapper import DataWrapper
-from .viewer._viewer import NDViewer
 
 __all__ = ["NDViewer", "DataWrapper", "imshow", "data"]
 
@@ -23,7 +23,7 @@ if TYPE_CHECKING:
     # these may be used externally, but are not guaranteed to be available at runtime
     # they must be used inside a TYPE_CHECKING block
 
-    from .viewer._dims_slider import DimKey as DimKey
-    from .viewer._dims_slider import Index as Index
-    from .viewer._dims_slider import Indices as Indices
-    from .viewer._dims_slider import Sizes as Sizes
+    from .views._qt._dims_slider import DimKey as DimKey
+    from .views._qt._dims_slider import Index as Index
+    from .views._qt._dims_slider import Indices as Indices
+    from .views._qt._dims_slider import Sizes as Sizes
