+++ src/ndv/controller/__init__.py
@@ -1,0 +1,3 @@
+from ._controller import ViewerController
+
+__all__ = ["ViewerController"]
