+++ src/ndv/models/__init__.py
@@ -1,6 +1,7 @@
 """Models for NDV."""
 
 from ._array_display_model import ArrayDisplayModel
+from ._data_display_model import DataDisplayModel
 from ._lut_model import LUTModel
 
-__all__ = ["ArrayDisplayModel", "LUTModel"]
+__all__ = ["ArrayDisplayModel", "LUTModel", "DataDisplayModel"]
