--- src/ndv/viewer/__init__.py
@@ -1,1 +1,0 @@
-"""viewer source."""
