+++ src/ndv/views/_qt/_components.py
