+++ src/ndv/views/_qt/_dims_slider.py
