+++ src/ndv/views/_qt/_save_button.py
@@ -7,7 +7,7 @@ from qtpy.QtWidgets import QFileDialog, QPushButton, QWidget
 from superqt.iconify import QIconifyIcon
 
 if TYPE_CHECKING:
-    from ._data_wrapper import DataWrapper
+    from ndv.models._data_wrapper import DataWrapper
 
 
 class SaveButton(QPushButton):
